feat: implement user actions for create, update, delete, and retrieve user
